requirements_telemark_abc.txt
streamlit
pandas
requests
python-dateutil
matplotlib
reportlab
